splitted tests, fixed some ids and fixed discover nav assertions
