+++ src/renderer/components/Onboarding/Screens/SelectUseCase/UseCaseOption.js
@@ -61,10 +61,10 @@ export function UseCaseOption({
   description,
   Illu,
   onClick,
-  id,
+  dataTestId,
 }: UseCaseOptionProps) {
   return (
-    <UseCaseOptionContainer id={id} onClick={onClick}>
+    <UseCaseOptionContainer data-test-id={dataTestId} onClick={onClick}>
       <Container>
         <Text
           mb="8px"
