+++ src/renderer/components/Onboarding/Screens/SelectUseCase/index.js
@@ -134,7 +134,7 @@ export function SelectUseCase({ sendEvent, context }: Props) {
           </LeftColumn>
           <RightColumn>
             <UseCaseOption
-              data-test-id="first-use"
+              dataTestId="onboarding-new-device"
               heading={t("onboarding.screens.selectUseCase.options.1.heading")}
               title={
                 <Trans
@@ -157,7 +157,7 @@ export function SelectUseCase({ sendEvent, context }: Props) {
           </LeftColumn>
           <RightColumn>
             <UseCaseOption
-              data-test-id="onboarding-initialized-device-step"
+              dataTestId="onboarding-initialized-device"
               heading={t("onboarding.screens.selectUseCase.options.2.heading")}
               title={
                 <Trans
@@ -172,7 +172,7 @@ export function SelectUseCase({ sendEvent, context }: Props) {
               onClick={() => onWrappedUseCase("CONNECT_SETUP_DEVICE")}
             />
             <UseCaseOption
-              id="restore-device"
+              dataTestId="onboarding-restore-device"
               heading={t("onboarding.screens.selectUseCase.options.3.heading")}
               title={t("onboarding.screens.selectUseCase.options.3.title")}
               description={
