+++ src/renderer/components/Onboarding/Screens/Tutorial/screens/PinCodeHowTo.js
@@ -178,7 +178,7 @@ export function PinCodeHowTo({ sendEvent, context }: Props) {
             <Trans i18nKey="onboarding.screens.tutorial.screens.pinCodeHowTo.buttons.prev" />
           </Text>
         </Button>
-        <Button id="pincode-howto-cta" primary onClick={onClickNext}>
+        <Button data-test-id="pincode-howto-cta" primary onClick={onClickNext}>
           <Text mr="12px" ff="Inter|Bold" fontSize={3} lineHeight="18px">
             <Trans i18nKey="onboarding.screens.tutorial.screens.pinCodeHowTo.buttons.next" />
           </Text>
