+++ tests/specs/layout.spec.ts
@@ -36,7 +36,7 @@ test("Layout", async ({ page }) => {
 
   await test.step("go to discover", async () => {
     await layout.goToDiscover();
-    await expect(page).toHaveURL(/.*\/platform\/.*/);
+    await expect(page).toHaveURL(/.*\/platform.*/);
     // FIXME: We need a way to ensure icons & images are rendered when we take a screenshot or retry
     // expect(await page.screenshot()).toMatchSnapshot("discover.png");
   });
